Rework C implementation
