+++ src/peer.c
@@ -14,8 +14,8 @@
 #include "peer.h"
 #include "message.h"
 #include "protocol.h"
-#include "secure_random.h"
-#include "secure_memory.h"
+#include "random.h"
+#include "memory.h"
 #include "validation.h"
 #include "scoring.h"
 #include "datadir.h"
